requirements.txt
torch>=1.8.0
transformers>=4.15.0
faiss-cpu>=1.7.0
whoosh>=2.7.4
scikit-learn>=0.24.0
tqdm>=4.60.0
jieba>=0.42.1
langchain-openai>=0.1.0
pyyaml>=6.0
modelscope>=1.4.0